Rename alert component
Used to be banner, alert is more accurate.
Rename class and methods.
