+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Pages/Admin/Trainings/List/Index.cshtml
@@ -26,7 +26,7 @@
 </div>
 }
 
-@if (TempData.HasGlobalBannerMessage())
+@if (TempData.HasGlobalAlertMessage())
 {
     <div class="o-container">
         <smart-spacer left="ExtraLarge" right="ExtraLarge" top="Small">
@@ -38,7 +38,7 @@
                 </div>
                 <div class="c-alert__body">
                     <div class="c-alert__text">
-                        <h4 class="c-h4">@TempData.GetGlobalBannerMessage().Message</h4>
+                        <h4 class="c-h4">@TempData.GetGlobalAlertMessage().Message</h4>
                         @*Show training's page link if it is published*@
                         @if(!string.IsNullOrEmpty(TempData["Url"]?.ToString())){
                             <div class="c-alert__message">
